Update I18n locale for Thai
Notice that Thai locale has to be in the Ruby file, as I need the interpolation for Buddhist year in those date and time. By trying to DRY up the code, I extracted those Active Record's error messages into a module and then use them later in the returning hash.