Copy nl translation file to nl-NL and nl-BE
Disclaimer: I am not a native speaker of this language. I have probably made a terrible mistake. Forgive me.

This is done because there are two groups of speakers in Netherlands: dutch (NL) and belgians (BE). Within Spree, we allow people to differentiate between the two languages because people have been willing and able to submit translations for both variants.

rails-i18n does not have these translations, and so it falls down when it attempts to ask for things like the number formatting rules (separators, delimiters, mainly).

You can work around this problem by copying nl.yml into your app's config/locale directory and calling it nl-NL.yml or nl-BE.yml if you wish, but that's the long way of going about it. So therefore I thought I'd do the copying within rails-i18n.

This will fix the problem that we're (@gidogeek and I, for now) seeing inside Spree.

Again, I am not a native speaker of this language. Patches welcome for any mistakes I have made.