Revert "Change about_x_* from 'cerca de' to 'hace'"
This reverts commit e6a1c4f1b081b193f45a7f011039f7e55779127e.

This commit broke the meaning of "distance_of_time_in_words". That
method is supposed to return the description of a time interval
without specifying if it's in the future or in the past. That is,
it should return "about 1 hour" instead of "about 1 hour ago" or
"in about 1 hour from now".

Well, this commit replaced "about 1 hour" with "1 hour ago", which is
wrong. You can't now use the "distance_of_time_in_words" method to
describe dates in the future, because the final result would be
something like "in about 2 hours ago from now".