Change confirmation error message  for es-* locales
This follows the changes introduced in rails/rails#5942, but does not include the attribute name in the message because it reads weird and it would require adding an article, which in turn would require knowing the grammatical gender of the attribute. E.g., "no coincide con *la* contraseña" or "no coincide con *el* número de cuenta".