The currency format and delimiter for PT locale (Portugal) is WRONG.
The typographic guide for Portugal is for example '1 000,00 €' and not '€1000.00'. You can verify this by consulting the official EU guide for Portugal:
http://publications.europa.eu/code/pt/pt-370303.htm

Attention: If you change the locale on the page http://publications.europa.eu/code/pt/pt-370303.htm to English (en), they will change the convention for UK convention which it's different that the Portugal convention (I don't know why they do that...).

Another simple way to verify this, if you have Mac OS X, go to Settings, Language and Region and change the region to Portugal or add Portuguese (Portugal) language. See the example that is returned. 

Please correct this. Thank you.