PT: Update Portuguese locales
Note that the negative (currency) format for Portuguese (Portugal) has
been changed to match the positive format (and the EU recommended
format).

Update translations for "long"/"short" to more natural forms
("longo"/"curto").  The feminine form of "long" ("longa") was selected
for the context of password_too_long, as "senha" is the most common
translation for "password".